chore: Remove unused CustomRepositoryHandler chore: Merge Repository and RepositoryHandler interfaces
