+++ src/main/java/org/terasology/launcher/packages/LocalRepository.java
@@ -18,6 +18,8 @@ package org.terasology.launcher.packages;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.terasology.launcher.packages.db.RepositoryConfiguration;
+import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
@@ -98,5 +100,10 @@ public class LocalRepository implements Repository {
     public Optional<Package> getPackage(PackageBuild pkgBuild, int version) {
         // TODO: Implement this
         return Optional.empty();
+    }
+
+    @Override
+    public List<Package> getPackageList(RepositoryConfiguration config) {
+        throw new NotImplementedException();
     }
 }
